// projects/evo-ui-kit/src/lib/components/evo-tooltip/enums/evo-tooltip-position.ts
export enum EvoTooltipPosition {
    TOP_START = 'top-start',
    TOP = 'top',
    TOP_END = 'top-end',
    BOTTOM_START = 'bottom-start',
    BOTTOM = 'bottom',
    BOTTOM_END = 'bottom-end',
    RIGHT_START = 'right-start',
    RIGHT = 'right',
    RIGHT_END = 'right-end',
    LEFT_START = 'left-start',
    LEFT = 'left',
    LEFT_END = 'left-end',
}